requirements.txt
requests
pandas
python-dotenv
psycopg2-binary
SQLAlchemy
pyarrow
notebook
numpy
scikit-learn
matplotlib
seaborn
wordcloud
plotly
nltk
textblob
fastapi
uvicorn[standard]
pydantic
streamlit
psutil
scipy
prometheus-fastapi-instrumentator
